Python_MiniGame_Fighter/requirements.txt
pyautogui
pillow
pynput
pygame